update blob index size (#457)

+++ inabox/tests/integration_test.go
@@ -204,7 +204,7 @@ func blobVerificationProofFromProto(verificationProof *disperserpb.BlobVerificat
 	}
 	return rollupbindings.EigenDARollupUtilsBlobVerificationProof{
 		BatchId:        verificationProof.GetBatchId(),
-		BlobIndex:      uint8(verificationProof.GetBlobIndex()),
+		BlobIndex:      verificationProof.GetBlobIndex(),
 		BatchMetadata:  batchMetadata,
 		InclusionProof: verificationProof.GetInclusionProof(),
 		QuorumIndices:  verificationProof.GetQuorumIndexes(),
